Loader: extract param resolution and resume warnings into params.rs
ResolvedParams, resolve_params, merge_no_flag, and resume_warnings all
revolve around merging CLI args with a checkpoint and don't belong in
main.rs. Pull them into a dedicated module. main.rs drops from 729 to
485 lines.
